refactor: extract monitor.c from somewm.c
Move all monitor/output management into a dedicated module:
display lifecycle (createmon, cleanupmon, closemon, updatemons),
rendering (rendermon, gpureset), output configuration (outputmgr*,
powermgrsetmode, requestmonstate), and queries (dirtomon, focusmon,
xytomon).

14 functions and 6 global listener structs extracted.

somewm.c: 6053 -> 5299 lines (-754)
monitor.c: 809 lines (new)
